refactor to use shared test class
